Move Requests to DashboardRouter level and firebasify over there
