+++ website/src/components/index.js
@@ -1,5 +1,5 @@
 import ProfileInfo from './ProfileInfo/ProfileInfo';
-import { FirebaseDashboardHeader as DashboardHeader } from './DashboardHeader/DashboardHeader';
+import DashboardHeader from './DashboardHeader/DashboardHeader';
 import MaterialIcon from './MaterialIcon/MaterialIcon';
 import HeaderIcons from './HeaderIcons/HeaderIcons';
 import IconButton from './IconButton/IconButton';
